+++ rv/chunks/chunk.py
@@ -3,11 +3,15 @@ from struct import pack
 
 class Chunk(object):
 
+    chnm = None
+    has_chff = False
+
     def chunks(self):
         yield (b'CHNM', pack('<I', self.chnm))
         yield (b'CHDT', self.chdt())
-        yield (b'CHFF', self.chff())
-        yield (b'CHFR', self.chfr())
+        if self.has_chff:
+            yield (b'CHFF', self.chff())
+            yield (b'CHFR', self.chfr())
 
     def chdt(self):
         return b''